worker/requirements.txt
openai-whisper==20240930
redis==5.2.1
